removed unused static property 'supported_*_props'; made schemaorg inheritance a bit simpler
