+++ src/cffconvert/cff_1_0_x/schemaorg.py
@@ -1,4 +1,3 @@
-import json
 from cffconvert.cff_1_x_x.authors.schemaorg import SchemaorgAuthor
 from cffconvert.cff_1_x_x.schemaorg import SchemaorgObjectShared as Shared
 from cffconvert.cff_1_x_x.urls.schemaorg import SchemaorgUrls
@@ -11,9 +10,6 @@ class SchemaorgObject(Shared):
         "1.0.2",
         "1.0.3"
     ]
-    supported_schemaorg_props = Shared.supported_schemaorg_props + [
-        "@context"
-    ]
 
     def __init__(self, cffobj, context="https://schema.org", initialize_empty=False):
         super().__init__(cffobj)
@@ -25,36 +21,6 @@ class SchemaorgObject(Shared):
             self.check_cffobj()
             self.add_all()
 
-    def __str__(self, sort_keys=True, indent=2):
-        data = {
-            "@context": self.context,
-            "@type": "SoftwareSourceCode",
-            "author": self.author,
-            "codeRepository": self.code_repository,
-            "datePublished": self.date_published,
-            "description": self.description,
-            "identifier": self.identifier,
-            "keywords": self.keywords,
-            "license": self.license,
-            "name": self.name,
-            "url": self.url,
-            "version": self.version
-        }
-        filtered = [item for item in data.items() if item[1] is not None]
-        return json.dumps(dict(filtered), sort_keys=sort_keys, indent=indent, ensure_ascii=False) + "\n"
-
-    def add_all(self):
-        self.add_author()           \
-            .add_date_published()   \
-            .add_description()      \
-            .add_identifier()       \
-            .add_keywords()         \
-            .add_license()          \
-            .add_name()             \
-            .add_urls()             \
-            .add_version()
-        return self
-
     def add_author(self):
         authors_cff = self.cffobj.get("authors", [])
         authors_schemaorg = [SchemaorgAuthor(a).as_dict() for a in authors_cff]
@@ -63,7 +29,7 @@ class SchemaorgObject(Shared):
 
     def add_contributor(self):
         # CFF v1.0.x doesn't have contributors
-        pass
+        return self
 
     def add_date_published(self):
         if "date-released" in self.cffobj.keys():
@@ -76,6 +42,11 @@ class SchemaorgObject(Shared):
     def add_identifier(self):
         if "doi" in self.cffobj.keys():
             self.identifier = f"https://doi.org/{self.cffobj['doi']}"
+        return self
+
+    def add_type(self):
+        # CFF v1.0.x assumes the metadata describes software
+        self.type = "SoftwareSourceCode"
         return self
 
     def add_urls(self):
