+++ src/cffconvert/cff_1_1_x/schemaorg.py
@@ -1,4 +1,3 @@
-import json
 from cffconvert.cff_1_x_x.authors.schemaorg import SchemaorgAuthor
 from cffconvert.cff_1_x_x.schemaorg import SchemaorgObjectShared as Shared
 from cffconvert.cff_1_x_x.urls.schemaorg import SchemaorgUrls
@@ -8,9 +7,6 @@ class SchemaorgObject(Shared):
 
     supported_cff_versions = [
         "1.1.0"
-    ]
-    supported_schemaorg_props = Shared.supported_schemaorg_props + [
-        "@context"
     ]
 
     def __init__(self, cffobj, context="https://schema.org", initialize_empty=False):
@@ -23,36 +19,6 @@ class SchemaorgObject(Shared):
             self.check_cffobj()
             self.add_all()
 
-    def __str__(self, sort_keys=True, indent=2):
-        data = {
-            "@context": self.context,
-            "@type": "SoftwareSourceCode",
-            "author": self.author,
-            "codeRepository": self.code_repository,
-            "datePublished": self.date_published,
-            "description": self.description,
-            "identifier": self.identifier,
-            "keywords": self.keywords,
-            "license": self.license,
-            "name": self.name,
-            "url": self.url,
-            "version": self.version
-        }
-        filtered = [item for item in data.items() if item[1] is not None]
-        return json.dumps(dict(filtered), sort_keys=sort_keys, indent=indent, ensure_ascii=False) + "\n"
-
-    def add_all(self):
-        self.add_author()           \
-            .add_date_published()   \
-            .add_description()      \
-            .add_identifier()       \
-            .add_keywords()         \
-            .add_license()          \
-            .add_name()             \
-            .add_urls()             \
-            .add_version()
-        return self
-
     def add_author(self):
         authors_cff = self.cffobj.get("authors", [])
         authors_schemaorg = [SchemaorgAuthor(a).as_dict() for a in authors_cff]
@@ -61,7 +27,7 @@ class SchemaorgObject(Shared):
 
     def add_contributor(self):
         # CFF v1.1.x doesn't have contributors
-        pass
+        return self
 
     def add_date_published(self):
         if "date-released" in self.cffobj.keys():
@@ -80,6 +46,11 @@ class SchemaorgObject(Shared):
                 if identifier["type"] == "doi":
                     self.identifier = f"https://doi.org/{identifier['value']}"
                     break
+        return self
+
+    def add_type(self):
+        # CFF v1.1.x assumes the metadata describes software
+        self.type = "SoftwareSourceCode"
         return self
 
     def add_urls(self):
