+++ src/cffconvert/cff_1_3_x/schemaorg.py
@@ -1,4 +1,3 @@
-import json
 from cffconvert.cff_1_x_x.authors.schemaorg import SchemaorgAuthor
 from cffconvert.cff_1_x_x.schemaorg import SchemaorgObjectShared as Shared
 from cffconvert.cff_1_x_x.urls.schemaorg import SchemaorgUrls
@@ -9,56 +8,16 @@ class SchemaorgObject(Shared):
     supported_cff_versions = [
         "1.3.0"
     ]
-    supported_schemaorg_props = Shared.supported_schemaorg_props + [
-        "@context",
-        "@type",
-        "contributor"
-    ]
 
     def __init__(self, cffobj, context="https://schema.org", initialize_empty=False):
         super().__init__(cffobj)
         self.context = context
-        self.contributor = None
-        self.type = None
         if initialize_empty:
             # clause for testing purposes
             pass
         else:
             self.check_cffobj()
             self.add_all()
-
-    def __str__(self, sort_keys=True, indent=2):
-        data = {
-            "@context": self.context,
-            "@type": self.type,
-            "author": self.author,
-            "contributor": self.contributor,
-            "codeRepository": self.code_repository,
-            "datePublished": self.date_published,
-            "description": self.description,
-            "identifier": self.identifier,
-            "keywords": self.keywords,
-            "license": self.license,
-            "name": self.name,
-            "url": self.url,
-            "version": self.version
-        }
-        filtered = [item for item in data.items() if item[1] is not None]
-        return json.dumps(dict(filtered), sort_keys=sort_keys, indent=indent, ensure_ascii=False) + "\n"
-
-    def add_all(self):
-        self.add_author()           \
-            .add_contributor()      \
-            .add_date_published()   \
-            .add_description()      \
-            .add_identifier()       \
-            .add_keywords()         \
-            .add_license()          \
-            .add_name()             \
-            .add_type()             \
-            .add_urls()             \
-            .add_version()
-        return self
 
     def add_author(self):
         authors_cff = self.cffobj.get("authors", [])
